[libc++] make atomic_waitable_traits in its own header
