+++ libcxx/include/atomic
@@ -607,6 +607,7 @@ template <class T>
 #  include <__atomic/atomic_flag.h>
 #  include <__atomic/atomic_init.h>
 #  include <__atomic/atomic_lock_free.h>
+#  include <__atomic/atomic_waitable_traits.h>
 #  include <__atomic/atomic_sync.h>
 #  include <__atomic/check_memory_order.h>
 #  include <__atomic/contention_t.h>
